Updated logic for get_updated_at to be up-to-date and cleaner
In GitHub's V3 API, there is no such field commit[comitter][date],
as the commit[comitter] field is metadata about the comitter. The
date that the last commit was created can now always be found at
commit[commit][comitter][date]. The logic for returning the last
modified time is changed to reflect this.

Also, the logic for determining if the pull request has been mod-
ified since the last bot action is simplified.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>