Re-implement failed job explanation logic
This commit consists of a couple of actions:

 - Revert "Revert "Implemented failed job explanation logic""
This reverts commit 9c4203c2b747c71cb6ee4777d1de364991c9cc61.

 - Added whitespace to valid prefix end characters
In order to keep the prefix detection logic backwards
compatible with the original literal space, the white-
space character class is added to the list of charac-
ters that are allowed to terminate the prefix.

 - Linked correct build url in flake response comment
In order to make it clear which Jenkins job URL is
expected by the robot from users explaining their
flakes, the correct URL is now linked from the bot
comment explaining the syntax.

 - Added link to repo, moved issue link in flake comment
In order to be clearer and more useful for contributors,
the text identifying the repository now links to the
repository page on GitHub, and the text identifying the
flake label links to the issue query showing all issues
with the flake label specified.

 - Added support for #123 quick-links for GitHub issues
In order to make the UX of this feature nicer for devs,
addition to the full URLs.

 - Improved readability and terseness of robot comment
In order to make the UX of this feature nicer for devs,
the comment that is posted when a job will not be re-
triggered is re-written to be terser and to follow a
more methodical "If <case>, then <action>" syntax to
help developers understand what they need to do.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>